@@ -1,8 +1,9 @@
 name: youtube_parser
-version: 1.1.0
+version: 2.2.0
 description: A minimalist library that extracts IDs from all kinds of YouTube urls
-homepage: https://github.com/blausoft/youtube_parser
+homepage: https://github.com/jayjah/youtube_parser
 environment:
-  sdk: '>=1.12.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 dev_dependencies:
-  test: '^1.6.0'
+  test: any
+  pedantic: ^1.11.1
